Changed type signature of JEvent::GetEvent
Resolves confusion as to who is responsible for instantiating and configuring JEvents, and when this happens. Now the user is only expected to extend JObjects, and JANA is responsible for providing and configuring JEvents.

+++ src/plugins/JTest/TextEventRecord.h
@@ -37,7 +37,7 @@
 #include <JANA/JObject.h>
 
 struct TextEventRecord: public JObject {
-    int data;
+    size_t data;
 
 };
